+++ Project/client/src/components/profile/MyReviewElement.vue
@@ -48,6 +48,7 @@
 
 <script>
 import axios from "axios";
+import swal from 'sweetalert'
 import API from "@/common/drf.js";
 import { mapState } from 'vuex';
 
@@ -77,18 +78,43 @@ export default {
       return config;
     },
     deleteReview: function (reviewId) {
-      const headers = this.setToken();
-      axios({
-        url: API.URL + API.ROUTES.reviewDetail + reviewId,
-        method: "delete",
-        headers,
+      swal({
+        title: "게시글을 삭제하시겠습니까?",
+        icon: "warning",
+        buttons: {
+          cancel: "취소",
+          confirm: {
+            text: "확인",
+            className: "confirm-btn"
+          },
+        },
       })
-      .then(() => {
-        this.$emit("myReviewsUpdate");
+      .then((isDelete) => {
+        if (isDelete) {
+          const headers = this.setToken()
+          axios({
+            url: API.URL + API.ROUTES.reviewDetail + reviewId,
+            method: "delete",
+            headers,
+          })
+          .then(() => {
+            this.$emit("myReviewsUpdate")
+            swal({
+              title: "게시글이 삭제되었습니다.",
+              icon: "success",
+              buttons: {
+                confirm: {
+                  text: "확인",
+                  className: "confirm-btn"
+                },
+              },
+            })
+          })
+          .catch((err) => {
+            console.log(err)
+          })
+        }
       })
-      .catch((err) => {
-        console.log(err);
-      });
     },
     moveToReviewForm: function (reviewId) {
       this.$router.push({ name: "Update", params: { reviewId: reviewId, flag: this.flag, userId: this.userId  } });
